requirements.txt
# --- Core web stack ---
flask==3.0.3
flask-cors==5.0.0
werkzeug==3.1.3

# --- PDF + OCR processing ---
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13

# --- Vector DB / embeddings ---
faiss-cpu==1.9.0
fastembed==0.4.2
numpy==1.26.4

# --- Google Gemini + auth/env ---
google-generativeai==0.8.3
python-dotenv==1.0.1

# --- LangChain + Google GenAI integration ---
langchain-core==0.3.34
langchain-community==0.3.13
langchain-google-genai==2.0.7

# --- DOCX parsing ---
python-docx==1.1.2

# --- Templating / misc ---
jinja2>=3.0.3
setuptools>=65.5.0
